+++ src/App.css
@@ -1,38 +1,22 @@
 .App {
-  text-align: center;
+    display: flex;
+    min-height: 100vh;
+    background-color: #ffffff;
+    flex-direction: column;
+    padding: 10px;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+.App .main-container {
+    width: 375px;
+    height: 600px;
+    border: 1px solid #e6e6e6;
+    margin: auto auto;
+    border-radius: 10px;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+@media screen and (max-width: 375px) {
+    .App .main-container {
+        border: none;
+        margin: 0;
+    }
 }
