+++ src/components/Home/Home.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Home() {
+  return <div> The mobile screen view </div>;
+}
